RBAS

+++ lib/shoppage.dart
@@ -1,8 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'dart:convert';
+import 'package:shared_preferences/shared_preferences.dart';
 
 class ShopPage extends StatefulWidget {
+  const ShopPage({super.key});
+
   @override
   _ShopPageState createState() => _ShopPageState();
 }
@@ -11,13 +14,22 @@ class _ShopPageState extends State<ShopPage> {
   String selectedCategory = 'Whey Protein';
   List<dynamic> products = [];
   List<dynamic> filteredProducts = [];
-  bool isLoading = true; // Loading indicator
+  bool isLoading = true;
   TextEditingController searchController = TextEditingController();
+  String userRole = 'user'; // Default role
 
   @override
   void initState() {
     super.initState();
+    _loadUserRole();
     fetchProducts();
+  }
+
+  Future<void> _loadUserRole() async {
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    setState(() {
+      userRole = prefs.getString('user') ?? 'user';
+    });
   }
 
   // Fetch data from API with error handling and timeout
@@ -25,7 +37,7 @@ class _ShopPageState extends State<ShopPage> {
     try {
       var url = 'https://gym-management-2.onrender.com/products/';
       var response =
-      await http.get(Uri.parse(url)).timeout(Duration(seconds: 10));
+      await http.get(Uri.parse(url)).timeout(const Duration(seconds: 10));
 
       if (response.statusCode == 200) {
         var data = json.decode(response.body);
@@ -76,6 +88,10 @@ class _ShopPageState extends State<ShopPage> {
   // Create a new product
   Future<void> createProduct(String name, String price, String desc,
       String image, int stock, String reviews) async {
+    if (userRole != 'admin') {
+      _showErrorMessage('You do not have permission to create products.');
+      return;
+    }
     try {
       var url = 'https://gym-management-2.onrender.com/products/';
       var response = await http.post(
@@ -103,6 +119,10 @@ class _ShopPageState extends State<ShopPage> {
 
   // Delete a product
   Future<void> deleteProduct(String productId) async {
+    if (userRole != 'admin') {
+      _showErrorMessage('You do not have permission to delete products.');
+      return;
+    }
     try {
       var url = 'https://gym-management-2.onrender.com/products/$productId';
       var response = await http.delete(Uri.parse(url));
@@ -120,6 +140,10 @@ class _ShopPageState extends State<ShopPage> {
   // Update a product
   Future<void> updateProduct(String productId, String name, String price,
       String desc, String image, int stock, String reviews) async {
+    if (userRole != 'admin') {
+      _showErrorMessage('You do not have permission to update products.');
+      return;
+    }
     try {
       var url = 'https://gym-management-2.onrender.com/products/$productId';
       var response = await http.put(
@@ -152,8 +176,8 @@ class _ShopPageState extends State<ShopPage> {
 
     return Scaffold(
       appBar: AppBar(
-        backgroundColor: Color(0xFF00B2B2),
-        title: Text('SHOPS'),
+        backgroundColor: const Color(0xFF00B2B2),
+        title: const Text('SHOPS'),
         actions: [
           Padding(
             padding: EdgeInsets.all(screenWidth * 0.013),
@@ -166,49 +190,48 @@ class _ShopPageState extends State<ShopPage> {
           ),
         ],
       ),
-      backgroundColor: Color(0xFF00B2B2),
+      backgroundColor: const Color(0xFF00B2B2),
       body: isLoading
-          ? Center(child: CircularProgressIndicator())// Show message if no products
+          ? const Center(child: CircularProgressIndicator())
           : Padding(
         padding: EdgeInsets.all(screenWidth * 0.04),
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            // Search Bar
             TextField(
               controller: searchController,
               decoration: InputDecoration(
                 filled: true,
-                fillColor: Color(0x96D9D9D9),
-                prefixIcon: Icon(Icons.search),
+                fillColor: const Color(0x96D9D9D9),
+                prefixIcon: const Icon(Icons.search),
                 hintText: 'Search',
                 border: OutlineInputBorder(
                   borderRadius: BorderRadius.circular(screenWidth * 0.07),
                   borderSide: BorderSide.none,
                 ),
               ),
-              onChanged: filterProducts, // Filter products as user types
+              onChanged: filterProducts,
             ),
-            SizedBox(height: screenHeight * 0.02), // Add some space below the search bar
+            SizedBox(height: screenHeight * 0.02),
 
-            // Create Button
-            ElevatedButton(
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) => CreateProductPage(
-                      onCreate: (name, price, desc, image, stock, reviews) {
-                        createProduct(name, price, desc, image, stock, reviews);
-                      },
+            if (userRole == 'admin') ...[
+              ElevatedButton(
+                onPressed: () {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) => CreateProductPage(
+                        onCreate: (name, price, desc, image, stock, reviews) {
+                          createProduct(name, price, desc, image, stock, reviews);
+                        },
+                      ),
                     ),
-                  ),
-                );
-              },
-              child: Text('Create New Product'),
-            ),
+                  );
+                },
+                child: const Text('Create New Product'),
+              ),
+            ],
 
-            // Product Grid
             Expanded(
               child: GridView.builder(
                 gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
@@ -224,7 +247,10 @@ class _ShopPageState extends State<ShopPage> {
                       Navigator.push(
                         context,
                         MaterialPageRoute(
-                          builder: (context) => ProductDetailPage(product: filteredProducts[index]),
+                          builder: (context) => ProductDetailPage(
+                            product: filteredProducts[index],
+                            userRole: userRole,
+                          ),
                         ),
                       );
                     },
@@ -251,11 +277,11 @@ class _ShopPageState extends State<ShopPage> {
             decoration: ShapeDecoration(
               color: Colors.white,
               shape: RoundedRectangleBorder(
-                side: BorderSide(width: 1, color: Color(0xFF9736C1)),
+                side: const BorderSide(width: 1, color: Color(0xFF9736C1)),
                 borderRadius: BorderRadius.circular(screenWidth * 0.05),
               ),
               image: DecorationImage(
-                image: NetworkImage(product['image']), // Use NetworkImage for images from API
+                image: NetworkImage(product['image']),
                 fit: BoxFit.contain,
               ),
             ),
@@ -267,7 +293,7 @@ class _ShopPageState extends State<ShopPage> {
               height: screenHeight * 0.11,
               padding: EdgeInsets.all(screenWidth * 0.02),
               decoration: ShapeDecoration(
-                color: Color(0xFFD2FFE8),
+                color: const Color(0xFFD2FFE8),
                 shape: RoundedRectangleBorder(
                   borderRadius: BorderRadius.circular(screenWidth * 0.05),
                 ),
@@ -280,35 +306,39 @@ class _ShopPageState extends State<ShopPage> {
                       children: [
                         Text(
                           product['name'],
-                          style: TextStyle(fontSize: 14, fontWeight: FontWeight.bold),
+                          style: const TextStyle(fontSize: 14, fontWeight: FontWeight.bold),
                         ),
                         Text(
                           '\$${product['price']}',
-                          style: TextStyle(fontSize: 12),
+                          style: const TextStyle(fontSize: 12),
                         ),
                       ],
                     ),
                   ),
-                  IconButton(
-                    icon: Icon(Icons.delete),
-                    onPressed: () {
-                      deleteProduct(product['_id']);
-                    },
-                  ),
-                  IconButton(
-                    icon: Icon(Icons.edit),
-                    onPressed: () {
-                      updateProduct(
-                        product['_id'],
-                        'Updated Name',
-                        '34.99',
-                        'Updated description',
-                        product['image'],
-                        product['stock'],
-                        product['reviews'],
-                      );
-                    },
-                  ),
+                  if (userRole == 'admin') ...[
+                    IconButton(
+                      icon: const Icon(Icons.delete),
+                      onPressed: () {
+                        deleteProduct(product['_id']);
+                      },
+                    ),
+                    IconButton(
+                      icon: const Icon(Icons.edit),
+                      onPressed: () {
+                        Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                            builder: (context) => UpdateProductPage(
+                              product: product,
+                              onUpdate: (id, name, price, desc, image, stock, reviews) {
+                                updateProduct(id, name, price, desc, image, stock, reviews);
+                              },
+                            ),
+                          ),
+                        );
+                      },
+                    ),
+                  ],
                 ],
               ),
             ),
@@ -323,7 +353,7 @@ class _ShopPageState extends State<ShopPage> {
 class CreateProductPage extends StatelessWidget {
   final Function(String, String, String, String, int, String) onCreate;
 
-  CreateProductPage({required this.onCreate});
+  CreateProductPage({super.key, required this.onCreate});
 
   final TextEditingController nameController = TextEditingController();
   final TextEditingController priceController = TextEditingController();
@@ -336,7 +366,7 @@ class CreateProductPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Create New Product'),
+        title: const Text('Create New Product'),
       ),
       body: Padding(
         padding: const EdgeInsets.all(16.0),
@@ -344,31 +374,31 @@ class CreateProductPage extends StatelessWidget {
           children: [
             TextField(
               controller: nameController,
-              decoration: InputDecoration(labelText: 'Product Name'),
+              decoration: const InputDecoration(labelText: 'Product Name'),
             ),
             TextField(
               controller: priceController,
-              decoration: InputDecoration(labelText: 'Price'),
+              decoration: const InputDecoration(labelText: 'Price'),
               keyboardType: TextInputType.number,
             ),
             TextField(
               controller: descController,
-              decoration: InputDecoration(labelText: 'Description'),
+              decoration: const InputDecoration(labelText: 'Description'),
             ),
             TextField(
               controller: imageController,
-              decoration: InputDecoration(labelText: 'Image URL'),
+              decoration: const InputDecoration(labelText: 'Image URL'),
             ),
             TextField(
               controller: stockController,
-              decoration: InputDecoration(labelText: 'Stock Quantity'),
+              decoration: const InputDecoration(labelText: 'Stock Quantity'),
               keyboardType: TextInputType.number,
             ),
             TextField(
               controller: reviewsController,
-              decoration: InputDecoration(labelText: 'Reviews'),
+              decoration: const InputDecoration(labelText: 'Reviews'),
             ),
-            SizedBox(height: 20),
+            const SizedBox(height: 20),
             ElevatedButton(
               onPressed: () {
                 onCreate(
@@ -381,7 +411,7 @@ class CreateProductPage extends StatelessWidget {
                 );
                 Navigator.pop(context); // Close the form after submission
               },
-              child: Text('Create Product'),
+              child: const Text('Create Product'),
             ),
           ],
         ),
@@ -391,10 +421,91 @@ class CreateProductPage extends StatelessWidget {
 }
 
 // Product Detail Page
+
+class UpdateProductPage extends StatelessWidget {
+  final Map<String, dynamic> product;
+  final Function(String, String, String, String, String, int, String) onUpdate;
+
+  UpdateProductPage({super.key, required this.product, required this.onUpdate});
+
+  final TextEditingController nameController = TextEditingController();
+  final TextEditingController priceController = TextEditingController();
+  final TextEditingController descController = TextEditingController();
+  final TextEditingController imageController = TextEditingController();
+  final TextEditingController stockController = TextEditingController();
+  final TextEditingController reviewsController = TextEditingController();
+
+  @override
+  Widget build(BuildContext context) {
+    nameController.text = product['name'];
+    priceController.text = product['price'].toString();
+    descController.text = product['desc'];
+    imageController.text = product['image'];
+    stockController.text = product['stock'].toString();
+    reviewsController.text = product['reviews'];
+
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Update Product'),
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          children: [
+            TextField(
+              controller: nameController,
+              decoration: const InputDecoration(labelText: 'Product Name'),
+            ),
+            TextField(
+              controller: priceController,
+              decoration: const InputDecoration(labelText: 'Price'),
+              keyboardType: TextInputType.number,
+            ),
+            TextField(
+              controller: descController,
+              decoration: const InputDecoration(labelText: 'Description'),
+            ),
+            TextField(
+              controller: imageController,
+              decoration: const InputDecoration(labelText: 'Image URL'),
+            ),
+            TextField(
+              controller: stockController,
+              decoration: const InputDecoration(labelText: 'Stock Quantity'),
+              keyboardType: TextInputType.number,
+            ),
+            TextField(
+              controller: reviewsController,
+              decoration: const InputDecoration(labelText: 'Reviews'),
+            ),
+            const SizedBox(height: 20),
+            ElevatedButton(
+              onPressed: () {
+                onUpdate(
+                  product['_id'],
+                  nameController.text,
+                  priceController.text,
+                  descController.text,
+                  imageController.text,
+                  int.parse(stockController.text),
+                  reviewsController.text,
+                );
+                Navigator.pop(context);
+              },
+              child: const Text('Update Product'),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
 class ProductDetailPage extends StatelessWidget {
   final Map<String, dynamic> product;
+  final String userRole;
 
-  ProductDetailPage({required this.product});
+  const ProductDetailPage({super.key, required this.product, required this.userRole});
 
   @override
   Widget build(BuildContext context) {
@@ -408,23 +519,33 @@ class ProductDetailPage extends StatelessWidget {
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
             Image.network(product['image']),
-            SizedBox(height: 16.0),
+            const SizedBox(height: 16.0),
             Text(
               product['desc'],
-              style: TextStyle(fontSize: 18.0),
+              style: const TextStyle(fontSize: 18.0),
             ),
-            SizedBox(height: 16.0),
+            const SizedBox(height: 16.0),
             Text(
               '\$${product['price']}',
-              style: TextStyle(fontSize: 24.0, fontWeight: FontWeight.bold),
+              style: const TextStyle(fontSize: 24.0, fontWeight: FontWeight.bold),
             ),
-            SizedBox(height: 16.0),
+            const SizedBox(height: 16.0),
             Text('Stock: ${product['stock']}'),
-            SizedBox(height: 16.0),
+            const SizedBox(height: 16.0),
             Text('Reviews: ${product['reviews']}'),
+            if (userRole != 'admin') ...[
+              const SizedBox(height: 16.0),
+              ElevatedButton(
+                onPressed: () {
+                  // Implement add to cart functionality
+                },
+                child: const Text('Add to Cart'),
+              ),
+            ],
           ],
         ),
       ),
     );
   }
 }
+
